Fixed the issues with Advanced Workbench Automation.
Reworked it to contain an internal buffer, it was having issues with
eating items if the crafting recipe created more than one item. It now
has an item overflow internally, when the recipe is changed, or the
block is broken, the overflow items are dropped on the ground so they
are not lost. The item overflow is used before crafting a new item.